getAll from products implemented, save from products implemented

+++ src/main/java/n2exercici1/Product.java
@@ -18,7 +18,7 @@ public abstract class Product{
 	}
 	
 	//---GETTERS & SETTERS---
-	public double getPrice() {
+	public float getPrice() {
 		return price;
 	}
